Bepaalde Errors gefixed

+++ src/drone/DroneObject.java
@@ -2,7 +2,7 @@ package drone;
 
 public interface DroneObject {
 	
-	public double getMass();
+	public float getMass();
 	
 	public float[] getGraviation();
